auth/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.7
cryptography==45.0.3
dnspython==2.7.0
docker==7.1.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.5
gherkin-official==29.0.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
parse==1.20.2
parse_type==0.6.4
passlib==1.7.4
pluggy==1.6.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
Pygments==2.19.1
pytest==8.4.0
pytest-bdd==8.1.0
pytest-mock==3.14.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.6
requests==2.32.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
testcontainers==4.10.0
typing_extensions==4.12.2
urllib3==2.4.0
uvicorn==0.32.1
wrapt==1.17.2
